diff: return diff_filepair from diff queue helpers
The `diff_addremove()` and `diff_change()` functions setup and queue
diffs, but do not return the `diff_filepair` added to the queue. In a
subsequent commit, modifications to `diff_filepair` need to take place
in certain cases after being queued.

Split out the queuing operations into `diff_filepair_addremove()` and
`diff_filepair_change()` which also return a handle to the queued
`diff_filepair`.

Signed-off-by: Justin Tobler <jltobler@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>